fix #7: 가게 평점 데이터타입 수정

+++ src/main/java/com/sparta/gourmate/domain/store/dto/StoreResponseDto.java
@@ -10,7 +10,7 @@ public class StoreResponseDto {
     private UUID categoryId;
     private String storeName;
     private String location;
-    private int averageRating;
+    private float averageRating;
 
 
     public StoreResponseDto(Store store) {
